requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
qasync==0.27.1
QtAwesome==1.3.1
QtPy==2.4.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.11.1
